backend/requirements.txt
# Core Web Backend
fastapi==0.115.14
uvicorn==0.35.0
python-dotenv==1.1.1
pydantic==2.11.7
starlette==0.46.2

# Async and Networking
httpx==0.27.0
aiohttp==3.9.5
anyio==4.2.0
sniffio==1.3.1
uvloop==0.19.0
watchfiles==1.1.0
httptools==0.6.1
websockets==11.0.3
h11==0.14.0

# Databases
sqlmodel==0.0.16
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pymongo==4.7.1

# Web3
web3==6.18.0
eth-account==0.11.0

# NLP / Utility
dateparser==1.2.0
requests==2.31.0

# Machine Learning / Torch
torch==2.2.2
numpy==1.26.4

motor

# Misc
click==8.1.7
PyYAML==6.0.1
typing-inspection==0.4.1
typing_extensions==4.12.2  # ✅ Required by pydantic>=2.11.7

# LangChain ecosystem (updated for compatibility)
langchain>=0.3.0
langchain-core>=0.3.0,<0.4
langchain-community>=0.3.0
langchain-huggingface==0.1.0
langsmith==0.3.45

# Embeddings / FAISS
faiss-cpu
sentence-transformers==2.7.0
huggingface-hub==0.23.0

# IPFS
ipfshttpclient>=0.8.0a2

speechrecognition
pyaudio
pyttsx3